backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
asyncssh==2.18.0
arq==0.26.1
redis==5.2.1
cryptography==44.0.0
PyYAML==6.0.2
